Setup Mac framework build. Setup Mac framework unit tests. Fixed account manager. Added 'all accounts' methods to account manager. Fixed account manager unit tests. Fixed pre-processor conditions in update user avatar operation. Heeded more compiler warnings about stringWithFormat: usage.

+++ Snapper/Source/Shared/SNBaseUserOperation.m
@@ -218,7 +218,7 @@
     SNResponse* response = [[SNResponse alloc] init];
 
     SNMetadata* meta = [[SNMetadata alloc] init];
-    meta.errorId = [NSString stringWithFormat:@"%d", error.code];
+    meta.errorId = [NSString stringWithFormat:@"%ld", (long)error.code];
     meta.errorMessage = [error localizedDescription];
 
     response.metadata = meta;
